requirements.txt
GitPython>=2.1.11
matplotlib>=3.0.3
baytune>=0.2.4
fire>=0.1.3
joblib>=0.13.2
numpy>=1.16.1
jupyter-core>=4.4.0
Keras>=2.2.4
scikit-learn>=0.20.2
tensorflow>=1.12.0
tqdm>=4.31.1
pytest>=4.4.0
git+https://www.github.com/keras-team/keras-contrib.git#egg=keras-contrib